+++ examples/run.php
@@ -6,10 +6,11 @@ use Castor\Attribute\AsTask;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Output\OutputInterface;
 
+use function Castor\capture;
 use function Castor\run;
 
 #[AsTask(description: 'Run a sub-process and display information about it')]
-function run_ls()
+function ls()
 {
     $process = run('ls -alh', quiet: true);
 
@@ -19,8 +20,18 @@ function run_ls()
     echo "\n";
 }
 
+#[AsTask(description: 'Run a sub-process and display information about it, with capture() function')]
+function whoami()
+{
+    // Note: we don't run `whoami` here, because it would break the tests suite
+    // for each different users
+    $whoami = capture('echo whoami');
+
+    echo "Hello: {$whoami}\n";
+}
+
 #[AsTask(description: 'Run a sub-process and display information about it, with ProcessHelper')]
-function run_with_process_helper(Application $application, OutputInterface $output)
+function with_process_helper(Application $application, OutputInterface $output)
 {
     if (!$output->isVeryVerbose()) {
         $output->writeln('Re-run with -vv, -vvv to see the output of the process.');